Added an event buffer to hell_loop()
To be used by the application to access
events of that frame. It is on the stack of hell_loop for speed, and is safe to
access inside the application frame function. The main reason for it is its kind of
ugly to have the main application handle events outside of its frame function -
its much nicer to be able to look at the frame function and see everything the
application does in one place. But, subscriptions are nice for things like
window resizing, and library handlers. Basically any event handling that the
application shouldn't need to worry about.